# docs/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "Bucket": {
        "description": "Where a report goes next.",
        "enum": [
          "brief",
          "follow_up",
          "watch",
          "park"
        ],
        "title": "Bucket",
        "type": "string"
      },
      "Coverage": {
        "additionalProperties": false,
        "description": "coverage_<set>.json, GET /api/v1/coverage/{set}: both matrices and the legend.",
        "properties": {
          "credibility": {
            "description": "Active reports assessed against the set, by credibility digit.",
            "items": {
              "$ref": "#/components/schemas/CredibilityBar"
            },
            "title": "Credibility",
            "type": "array"
          },
          "lanes": {
            "$ref": "#/components/schemas/CoverageMatrixView",
            "description": "Requirements down, lanes across."
          },
          "legend": {
            "items": {
              "$ref": "#/components/schemas/LegendEntry"
            },
            "title": "Legend",
            "type": "array"
          },
          "set_id": {
            "title": "Set Id",
            "type": "string"
          },
          "set_name": {
            "title": "Set Name",
            "type": "string"
          },
          "taxonomy": {
            "description": "Taxonomy nodes down, one per axis.",
            "items": {
              "$ref": "#/components/schemas/CoverageAxisView"
            },
            "title": "Taxonomy",
            "type": "array"
          }
        },
        "required": [
          "set_id",
          "set_name",
          "legend",
          "lanes",
          "taxonomy",
          "credibility"
        ],
        "title": "Coverage",
        "type": "object"
      },
...
      "Credibility": {
        "description": "The Admiralty credibility of one report's information, 1 (best) to 6 (cannot be judged).\n\nThe analyst judges credibility per report; the source's reliability is\nfixed by the watchlist. The two together make the grading, such as B2.",
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "title": "Credibility",
        "type": "integer"
      },
...
      "Direction": {
        "description": "Which way a report bears on one requirement.",
        "enum": [
          "supports",
          "threatens",
          "neutral",
          "informs_baseline"
        ],
        "title": "Direction",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IndicatorState": {
        "description": "Where a report is in the indications and warnings lifecycle.",
        "enum": [
          "emerging",
          "tracking",
          "reinforced",
          "matured",
          "faded",
          "parked"
        ],
        "title": "IndicatorState",
        "type": "string"
      },
      "LegendEntry": {
        "additionalProperties": false,
        "description": "One coverage status in the legend.",
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "meaning": {
            "title": "Meaning",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CoverageStatus"
          }
        },
        "required": [
          "status",
          "label",
          "meaning"
        ],
        "title": "LegendEntry",
        "type": "object"
      },
...
      "Picture": {
        "additionalProperties": false,
        "description": "picture_<set>.json, GET /api/v1/requirement-sets/{id}/picture.",
        "properties": {
          "datecheck": {
            "$ref": "#/components/schemas/DateCheckCounts"
          },
          "impact_groups": {
            "items": {
              "$ref": "#/components/schemas/GroupTile"
            },
            "title": "Impact Groups",
            "type": "array"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementTile"
            },
            "title": "Requirements",
            "type": "array"
          },
          "set_id": {
            "title": "Set Id",
            "type": "string"
          },
          "set_name": {
            "title": "Set Name",
            "type": "string"
          },
          "today": {
            "description": "YYYY-MM-DD.",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}$",
            "title": "Today",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "wbo_groups": {
            "items": {
              "$ref": "#/components/schemas/GroupTile"
            },
            "title": "Wbo Groups",
            "type": "array"
          }
        },
        "required": [
          "set_id",
          "set_name",
          "version",
          "today",
          "wbo_groups",
          "impact_groups",
          "requirements",
          "datecheck"
        ],
        "title": "Picture",
        "type": "object"
      },
      "ProductDetail": {
        "additionalProperties": false,
        "description": "products/<id>.json, GET /api/v1/products/{id}: one product's sections and Markdown.",
        "properties": {
          "body": {
            "description": "The product as Markdown.",
            "title": "Body",
            "type": "string"
          },
          "generated_on": {
            "description": "YYYY-MM-DD.",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}$",
            "title": "Generated On",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lead": {
            "items": {
              "type": "string"
            },
            "title": "Lead",
            "type": "array"
          },
          "level": {
            "$ref": "#/components/schemas/ProductLevel"
          },
          "markdown_file": {
            "title": "Markdown File",
            "type": "string"
          },
          "markdown_path": {
            "description": "Where the Markdown lives, relative to the data dir.",
            "title": "Markdown Path",
            "type": "string"
          },
          "method_note": {
            "title": "Method Note",
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/PeriodModel"
          },
          "report_ids": {
            "items": {
              "type": "string"
            },
            "title": "Report Ids",
            "type": "array"
          },
          "requirement_set_id": {
            "title": "Requirement Set Id",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionModel"
            },
            "title": "Sections",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "level",
          "requirement_set_id",
          "period",
          "generated_on",
          "title",
          "lead",
          "sections",
          "report_ids",
          "method_note",
          "markdown_file",
          "markdown_path",
          "body"
        ],
        "title": "ProductDetail",
        "type": "object"
      },
...
      "Reliability": {
        "description": "The Admiralty reliability of a source, A (best) to F (cannot be judged).\n\nReliability belongs to the source, never to one report: the analyst judges\ncredibility (1 to 6) per report, and the two together make the grading.",
        "enum": [
          "A",
          "B",
          "C",
          "D",
          "E",
          "F"
        ],
        "title": "Reliability",
        "type": "string"
      },
...
      "RequirementSetDetail": {
        "additionalProperties": false,
        "description": "requirement_set_<id>.json, GET /api/v1/requirement-sets/{id}.",
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupView"
            },
            "title": "Groups",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementView"
            },
            "title": "Requirements",
            "type": "array"
          },
          "source_doc": {
            "title": "Source Doc",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "source_doc",
          "is_default",
          "requirements",
          "groups"
        ],
        "title": "RequirementSetDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only development and analyst API serving the same JSON documents as the published snapshot.",
    "title": "Gwylio read API",
    "version": "0.1.0",
    "x-generated": "Generated by `gwylio schema` from the read API in backend/src/gwylio/api/app.py. Do not edit by hand."
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/coverage/{set_id}": {
      "get": {
        "description": "coverage_<set>.json: requirements by lane and taxonomy nodes by count, with the legend.",
        "operationId": "get_coverage_api_v1_coverage__set_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "title": "Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coverage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The coverage audit"
      }
    },
...
    "/api/v1/products/{product_id}": {
      "get": {
        "description": "products/<id>.json: one product's sections and its Markdown.",
        "operationId": "get_product_api_v1_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One product"
      }
    },
    "/api/v1/reports": {
      "get": {
        "description": "reports.json: every report, by id; with filters, only those that pass every one.",
        "operationId": "get_reports_api_v1_reports_get",
        "parameters": [
          {
            "description": "Assessed against this requirement set.",
            "in": "query",
            "name": "set",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Assessed against this requirement set.",
              "title": "Set"
            }
          },
          {
            "description": "Assessed against this requirement id.",
            "in": "query",
            "name": "requirement",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Assessed against this requirement id.",
              "title": "Requirement"
            }
          },
          {
            "description": "An assessment in this direction (on the requirement or set if given).",
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Direction"
                },
                {
                  "type": "null"
                }
              ],
              "description": "An assessment in this direction (on the requirement or set if given).",
              "title": "Direction"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IndicatorState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Bucket"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bucket"
            }
          },
          {
            "in": "query",
            "name": "lane",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lane"
            }
          },
          {
            "in": "query",
            "name": "hazard",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hazard"
            }
          },
          {
            "in": "query",
            "name": "place",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Place"
            }
          },
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          },
          {
            "description": "Latest history entry on or after this date.",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Latest history entry on or after this date.",
              "title": "Since"
            }
          },
          {
            "description": "Text in the title or summary.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text in the title or summary.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "reliability",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Reliability"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reliability"
            }
          },
          {
            "in": "query",
            "name": "credibility",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Credibility"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Credibility"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportSummary"
                  },
                  "title": "Response Get Reports Api V1 Reports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The register, filtered"
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "description": "reports/<id>.json: one report with assessments, tags, sightings and history.",
        "operationId": "get_report_api_v1_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One report"
      }
    },
...
    "/api/v1/requirement-sets/{set_id}": {
      "get": {
        "description": "requirement_set_<id>.json: one requirement set with requirements and groups.",
        "operationId": "get_requirement_set_api_v1_requirement_sets__set_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "title": "Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementSetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One requirement set"
      }
    },
    "/api/v1/requirement-sets/{set_id}/picture": {
      "get": {
        "description": "picture_<set>.json: tiles per group and requirement, and the date check counts.",
        "operationId": "get_picture_api_v1_requirement_sets__set_id__picture_get",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "title": "Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Picture"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One requirement set's picture"
      }
    },
...
    "/api/v1/scan-runs/{run_id}": {
      "get": {
        "description": "runs/<id>.json: one run's funnel, shares, dispositions and credibility spread.",
        "operationId": "get_run_api_v1_scan_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One scan run"
      }
    },
...
    "/api/v1/sources/{source_id}/yield": {
      "get": {
        "description": "One watched source with its yield (the matching entry of sources.json).",
        "operationId": "get_source_yield_api_v1_sources__source_id__yield_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One source's yield"
      }
    }
  }
}